identity: Restructure remaining v2 extensions code
Model things after how we do it for all other modules.
